+++ app/src/main/java/com/paulik/professionaldevelopment/ui/favorite/FavoriteWordViewModel.kt
@@ -25,7 +25,6 @@ class FavoriteWordViewModel(
         Dispatchers.Main
                 + SupervisorJob()
                 + CoroutineExceptionHandler { _, throwable ->
-            // todo
         })
 
     private fun cancelJob() {
