+++ app/src/main/java/com/paulik/professionaldevelopment/ui/favorite/FavoritesWordsFragment.kt
@@ -107,9 +107,4 @@ class FavoritesWordsFragment : ViewBindingFragment<FragmentFavoritesWordsBinding
         super.onDetach()
         scopeFavorite.close()
     }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() = FavoritesWordsFragment()
-    }
 }
