+++ app/src/main/java/com/paulik/professionaldevelopment/ui/root/RootActivity.kt
@@ -1,12 +1,20 @@
 package com.paulik.professionaldevelopment.ui.root
 
+import android.app.Activity
+import android.content.Intent
 import android.os.Bundle
 import android.view.KeyEvent
 import android.view.View
-import android.view.ViewTreeObserver
-import androidx.core.splashscreen.SplashScreen
+import android.widget.Toast
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.fragment.app.Fragment
+import com.google.android.material.snackbar.Snackbar
+import com.google.android.play.core.appupdate.AppUpdateManager
+import com.google.android.play.core.appupdate.AppUpdateManagerFactory
+import com.google.android.play.core.install.InstallStateUpdatedListener
+import com.google.android.play.core.install.model.AppUpdateType
+import com.google.android.play.core.install.model.InstallStatus
+import com.google.android.play.core.install.model.UpdateAvailability
 import com.paulik.core.ViewBindingActivity
 import com.paulik.historyscreen.HistoryWordTranslationFragment
 import com.paulik.professionaldevelopment.R
@@ -18,6 +26,7 @@ import com.paulik.professionaldevelopment.ui.translation.WordTranslationFragment
 import com.paulik.professionaldevelopment.ui.translation.descriptios.DescriptionWordTranslationFragment
 
 private const val TAG_ROOT_CONTAINER_LAYOUT_KEY = "TAG_ROOT_CONTAINER_LAYOUT_KEY"
+private const val REQUEST_CODE_CONNECTIVITY_SETTINGS = 42
 
 class RootActivity : ViewBindingActivity<ActivityRootBinding>(
     ActivityRootBinding::inflate
@@ -27,6 +36,8 @@ class RootActivity : ViewBindingActivity<ActivityRootBinding>(
     HistoryWordTranslationFragment.Controller,
     FavoritesWordsFragment.Controller {
 
+    private lateinit var appUpdateManager: AppUpdateManager
+
     private val favoritesWordsFragment: FavoritesWordsFragment by lazy {
         FavoritesWordsFragment()
     }
@@ -34,8 +45,9 @@ class RootActivity : ViewBindingActivity<ActivityRootBinding>(
 
     private var contentHasLoaded = true
     override fun onCreate(savedInstanceState: Bundle?) {
-        val splashScreen = installSplashScreen()
+        installSplashScreen()
         super.onCreate(savedInstanceState)
+        appUpdateManager = AppUpdateManagerFactory.create(this)
 
         onBottomNaviBar()
 
@@ -44,21 +56,6 @@ class RootActivity : ViewBindingActivity<ActivityRootBinding>(
         } else {
             // todo другое
         }
-        setupSplashScreen(splashScreen)
-    }
-
-    private fun setupSplashScreen(splashScreen: SplashScreen) {
-        val content: View = findViewById(android.R.id.content)
-        content.viewTreeObserver.addOnPreDrawListener(
-            object : ViewTreeObserver.OnPreDrawListener {
-                override fun onPreDraw(): Boolean {
-                    return if (contentHasLoaded) {
-                        content.viewTreeObserver.removeOnPreDrawListener(this)
-                        true
-                    } else false
-                }
-            }
-        )
     }
 
     private fun onBottomNaviBar() {
@@ -167,5 +164,63 @@ class RootActivity : ViewBindingActivity<ActivityRootBinding>(
             return true
         }
         return super.onKeyDown(keyCode, event)
+    }
+
+    private val stateUpdatedListener: InstallStateUpdatedListener =
+        InstallStateUpdatedListener { state ->
+            state.let {
+                if (it.installStatus() == InstallStatus.DOWNLOADED) {
+                    popupSnackbarForCompleteUpdate()
+                }
+            }
+        }
+
+    private fun popupSnackbarForCompleteUpdate() {
+        val rootView = findViewById<View>(android.R.id.content)
+        Snackbar.make(
+            rootView,
+            "Загружено обновление",
+            Snackbar.LENGTH_INDEFINITE
+        ).apply {
+            setAction("RESTART") { appUpdateManager.completeUpdate() }
+            show()
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        appUpdateManager
+            .appUpdateInfo
+            .addOnSuccessListener { appUpdateInfo ->
+                if (appUpdateInfo.installStatus() == InstallStatus.DOWNLOADED) {
+                    popupSnackbarForCompleteUpdate()
+                }
+                if (appUpdateInfo.updateAvailability()
+                    == UpdateAvailability.DEVELOPER_TRIGGERED_UPDATE_IN_PROGRESS
+                ) {
+                    appUpdateManager.startUpdateFlowForResult(
+                        appUpdateInfo,
+                        AppUpdateType.IMMEDIATE,
+                        this,
+                        REQUEST_CODE_CONNECTIVITY_SETTINGS
+                    )
+                }
+            }
+    }
+
+    @Deprecated("Deprecated in Java")
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode == REQUEST_CODE_CONNECTIVITY_SETTINGS) {
+            if (requestCode == Activity.RESULT_OK) {
+                appUpdateManager.unregisterListener(stateUpdatedListener)
+                val result = data?.getStringExtra("result")
+                Toast.makeText(this, "Результат: $result", Toast.LENGTH_SHORT).show()
+            } else {
+                Toast.makeText(this, "Result code: $resultCode", Toast.LENGTH_SHORT)
+                    .show()
+            }
+        }
     }
 }
