+++ app/src/main/java/com/paulik/professionaldevelopment/ui/translation/WordTranslationFragment.kt
@@ -1,6 +1,5 @@
 package com.paulik.professionaldevelopment.ui.translation
 
-import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Build
@@ -14,18 +13,10 @@ import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.android.material.snackbar.Snackbar
-import com.google.android.play.core.appupdate.AppUpdateManager
-import com.google.android.play.core.appupdate.AppUpdateManagerFactory
-import com.google.android.play.core.install.InstallStateUpdatedListener
-import com.google.android.play.core.install.model.AppUpdateType
-import com.google.android.play.core.install.model.InstallStatus
-import com.google.android.play.core.install.model.UpdateAvailability
 import com.paulik.core.ViewBindingWordTranslationFragment
 import com.paulik.models.entity.DataEntity
 import com.paulik.professionaldevelopment.R
 import com.paulik.professionaldevelopment.databinding.FragmentWordTranslationBinding
-import com.paulik.professionaldevelopment.ui.root.RootActivity
 import com.paulik.professionaldevelopment.ui.translation.dialog.SearchDialogFragment
 import com.paulik.repository.convertMeaningsToString
 import com.paulik.utils.network.isOnline
@@ -42,8 +33,6 @@ class WordTranslationFragment : ViewBindingWordTranslationFragment<FragmentWordT
 
     private var word: String? = null
     private var flagHistory: Boolean = false
-
-    private lateinit var appUpdateManager: AppUpdateManager
 
     override lateinit var viewModel: WordTranslationViewModel
 
@@ -99,26 +88,6 @@ class WordTranslationFragment : ViewBindingWordTranslationFragment<FragmentWordT
                 }
             }
         }
-    private val stateUpdatedListener: InstallStateUpdatedListener =
-        InstallStateUpdatedListener { state ->
-            state.let {
-                if (it.installStatus() == InstallStatus.DOWNLOADED) {
-                    popupSnackbarForCompleteUpdate()
-                }
-            }
-        }
-
-    private fun popupSnackbarForCompleteUpdate() {
-        val rootView = requireView()
-        Snackbar.make(
-            rootView,
-            "Загружено обновление",
-            Snackbar.LENGTH_INDEFINITE
-        ).apply {
-            setAction("RESTART") { appUpdateManager.completeUpdate() }
-            show()
-        }
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -128,8 +97,6 @@ class WordTranslationFragment : ViewBindingWordTranslationFragment<FragmentWordT
             flagHistory = it.getBoolean(FLAG_HISTORY_KEY)
         }
 
-        appUpdateManager = AppUpdateManagerFactory.create(requireContext())
-
         setHasOptionsMenu(true)
         initViewModel()
         initViews()
@@ -137,43 +104,6 @@ class WordTranslationFragment : ViewBindingWordTranslationFragment<FragmentWordT
 
     override fun setDataToAdapter(data: List<DataEntity>) {
         adapter.setData(data)
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        appUpdateManager
-            .appUpdateInfo
-            .addOnSuccessListener { appUpdateInfo ->
-                if (appUpdateInfo.installStatus() == InstallStatus.DOWNLOADED) {
-                    popupSnackbarForCompleteUpdate()
-                }
-                if (appUpdateInfo.updateAvailability()
-                    == UpdateAvailability.DEVELOPER_TRIGGERED_UPDATE_IN_PROGRESS
-                ) {
-                    (activity as WordTranslationFragment).appUpdateManager.startUpdateFlowForResult(
-                        appUpdateInfo,
-                        AppUpdateType.IMMEDIATE,
-                        activity as RootActivity,
-                        REQUEST_CODE_CONNECTIVITY_SETTINGS
-                    )
-                }
-            }
-    }
-
-    @Deprecated("Deprecated in Java")
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == REQUEST_CODE_CONNECTIVITY_SETTINGS) {
-            if (requestCode == Activity.RESULT_OK) {
-                appUpdateManager.unregisterListener(stateUpdatedListener)
-                val result = data?.getStringExtra("result")
-                Toast.makeText(requireContext(), "Результат: $result", Toast.LENGTH_SHORT).show()
-            } else {
-                Toast.makeText(requireContext(), "Result code: $resultCode", Toast.LENGTH_SHORT)
-                    .show()
-            }
-        }
     }
 
     @Deprecated("Deprecated in Java")
