+++ app/src/main/res/values-v31/styles.xml
@@ -2,9 +2,9 @@
 <resources>
     <!-- Splash Screen Theme (Android 12+) -->
     <style name="Theme.App.Starting" parent="Theme.SplashScreen.IconBackground">
-        <item name="windowSplashScreenBackground">@color/yellow6</item>
-        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_launcher_foreground</item>
-        <item name="windowSplashScreenIconBackgroundColor">@color/yellow6</item>
+        <item name="android:windowSplashScreenBackground">@color/yellow6</item>
+        <item name="android:windowSplashScreenAnimatedIcon">@drawable/ic_launcher_foreground</item>
+        <item name="android:windowSplashScreenIconBackgroundColor">@color/yellow6</item>
         <item name="android:windowSplashScreenBrandingImage">@drawable/label_android_12_new</item>
         <item name="postSplashScreenTheme">@style/AppTheme</item>
     </style>
